refactor(card menu): move into its own namespace folder
