+++ src/pages/Overview.js
@@ -5,7 +5,7 @@ import { Content as ContentLayout } from '../modules/layout/Content'
 import { SideBar as SideBarLayout } from '../modules/layout/SideBar'
 import { Content } from '../modules/Content'
 import { Sidebar } from '../modules/Sidebar'
-import { CardMenu } from '../modules/CardMenu'
+import { CardMenu } from '../modules/CardMenu/CardMenu'
 import { sectionOrder } from '../constants/sectionOrder'
 
 export const Overview = ({ name, match }) => {
